+++ src/components/FeaturePanel/FeatureImages/NoImage.tsx
@@ -3,7 +3,7 @@ import styled from '@emotion/styled';
 import { grey } from '@mui/material/colors';
 import { useFeatureContext } from '../../utils/FeatureContext';
 import { HEIGHT } from './helpers';
-import { icons } from '../../../assets/icons';
+import { iconsLookup } from '../../utils/icons/iconsLookup';
 
 const Wrapper = styled.div`
   position: relative;
@@ -28,7 +28,7 @@ export const NoImage = () => {
   const { feature } = useFeatureContext();
   const { properties } = feature;
   const ico = properties.class;
-  const icon = icons.includes(ico) ? ico : 'information';
+  const icon = iconsLookup.includes(ico) ? ico : 'information';
 
   return (
     <Wrapper>
