+++ src/components/utils/icons/PoiIcon.tsx
@@ -1,12 +1,12 @@
-import { useUserThemeContext } from '../../helpers/theme';
+import { useUserThemeContext } from '../../../helpers/theme';
 import { useTheme } from '@mui/material';
-import { AreaIcon } from '../FeaturePanel/Climbing/AreaIcon';
-import { CragIcon } from '../FeaturePanel/Climbing/CragIcon';
+import { AreaIcon } from '../../FeaturePanel/Climbing/AreaIcon';
+import { CragIcon } from '../../FeaturePanel/Climbing/CragIcon';
 import ShowChartIcon from '@mui/icons-material/ShowChart';
 import { Maki } from './Maki';
 import React from 'react';
 import styled from '@emotion/styled';
-import { FeatureTags } from '../../services/types';
+import { FeatureTags } from '../../../services/types';
 
 const Container = styled.span`
   margin-right: 6px;
